efreet: Do not recurse into icon dirs
There will be a lot of dirs below an icon dir, so it is a lot to watch
all. The icon theme spec says that the top level dir will be updated on
icon theme update, which it will on a standard linux system. A temporary
file will be created and deleted below the top level dir.

git-svn-id: http://svn.enlightenment.org/svn/e/trunk/efreet@79126 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33